TS: ProteinListFilters/TaxonomyFilter
